Refactor search command to split the files into mutliple files
This is to prepare for a future refactor across search and list commands that are almost identical.
